+++ .gitignore
@@ -26,3 +26,9 @@ gradle/wrapper
 
 #ignore changes to the apps properties file
 jdm.properties
+
+# Ignore Gradle project-specific cache directory
+.gradle
+
+# Ignore Gradle build output directory
+build
