+++ lib/src/main/java/edu/touro/mco152/bm/persist/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * A container for utility classes that help with handling persistent
+ * data especially to get statistical data about the device we are testing.
+ */
+package edu.touro.mco152.bm.persist;
