+++ lib/src/main/java/org/example/Library.java
@@ -1,0 +1,10 @@
+/*
+ * This source file was generated by the Gradle 'init' task
+ */
+package org.example;
+
+public class Library {
+    public boolean someLibraryMethod() {
+        return true;
+    }
+}
