--- src/edu/touro/mco152/bm/persist/package-info.java
@@ -1,5 +1,0 @@
-/**
- * A container for utility classes that help with handling persistent
- * data especially to get statistical data about the device we are testing.
- */
-package edu.touro.mco152.bm.persist;
