--- src/edu/touro/mco152/bm/ui/Gui.java
@@ -1,170 +1,0 @@
-package edu.touro.mco152.bm.ui;
-
-import edu.touro.mco152.bm.App;
-import edu.touro.mco152.bm.DiskMark;
-import org.jfree.chart.ChartFactory;
-import org.jfree.chart.ChartPanel;
-import org.jfree.chart.JFreeChart;
-import org.jfree.chart.axis.NumberAxis;
-import org.jfree.chart.plot.PlotOrientation;
-import org.jfree.chart.plot.XYPlot;
-import org.jfree.data.xy.XYSeries;
-import org.jfree.data.xy.XYSeriesCollection;
-
-import javax.swing.*;
-import java.awt.*;
-import java.io.Serial;
-import java.text.NumberFormat;
-
-/**
- * Creates and populates a graph with data from the current run, and
- * stores gui references for easy access.
- */
-public final class Gui {
-
-    public static ChartPanel chartPanel = null;
-    public static MainFrame mainFrame = null;
-    public static SelectFrame selFrame = null;
-    public static XYSeries wSeries, wAvgSeries, wMaxSeries, wMinSeries;
-    public static XYSeries rSeries, rAvgSeries, rMaxSeries, rMinSeries;
-    public static JFreeChart chart;
-    public static JProgressBar progressBar = null;
-    public static RunPanel runPanel = null;
-
-    /**
-     * Creates a ChartPanel for the GUI to display information related to
-     * operations like reading and writing for the current run.
-     * @return the ChartPanel instance to the MainFrame class.
-     * This method initializes the required XYSeries for writes and reads along with their average, maximum, and minimum values.
-     */
-    public static ChartPanel createChartPanel() {
-
-        wSeries = new XYSeries("Writes");
-        wAvgSeries = new XYSeries("Write Avg");
-        wMaxSeries = new XYSeries("Write Max");
-        wMinSeries = new XYSeries("Write Min");
-
-        rSeries = new XYSeries("Reads");
-        rAvgSeries = new XYSeries("Read Avg");
-        rMaxSeries = new XYSeries("Read Max");
-        rMinSeries = new XYSeries("Read Min");
-
-        XYSeriesCollection dataset = new XYSeriesCollection();
-        dataset.addSeries(wSeries);
-        dataset.addSeries(wAvgSeries);
-        dataset.addSeries(wMaxSeries);
-        dataset.addSeries(wMinSeries);
-        dataset.addSeries(rSeries);
-        dataset.addSeries(rAvgSeries);
-        dataset.addSeries(rMaxSeries);
-        dataset.addSeries(rMinSeries);
-
-        chart = ChartFactory.createXYLineChart(
-                "XY Chart", // Title
-                null, // x-axis Label
-                "Bandwidth MB/s", // y-axis Label
-                dataset, // Dataset
-                PlotOrientation.VERTICAL, // Plot Orientation
-                true,// Show Legend
-                true, // Use tooltips
-                false // Configure chart to generate URLs?
-        );
-        XYPlot plot = chart.getXYPlot();
-        plot.setBackgroundPaint(Color.DARK_GRAY);
-        ((NumberAxis) plot.getRangeAxis()).setAutoRangeIncludesZero(false);
-        NumberAxis range = (NumberAxis) plot.getDomainAxis();
-        range.setNumberFormatOverride(NumberFormat.getNumberInstance());
-        chart.getTitle().setVisible(false);
-        chartPanel = new ChartPanel(chart) {
-            // Only way to set the size of chart panel
-            // ref: http://www.jfree.org/phpBB2/viewtopic.php?p=75516
-            @Serial
-            private static final long serialVersionUID = 1L;
-
-            @Override
-            public Dimension getPreferredSize() {
-                return new Dimension(500, 325);
-            }
-        };
-
-        plot.getRenderer().setSeriesPaint(0, Color.YELLOW);
-        plot.getRenderer().setSeriesPaint(1, Color.WHITE);
-        plot.getRenderer().setSeriesPaint(2, Color.GREEN);
-        plot.getRenderer().setSeriesPaint(3, Color.RED);
-        plot.getRenderer().setSeriesPaint(4, Color.LIGHT_GRAY);
-        plot.getRenderer().setSeriesPaint(5, Color.ORANGE);
-        plot.getRenderer().setSeriesPaint(6, Color.GREEN);
-        plot.getRenderer().setSeriesPaint(7, Color.RED);
-        updateLegend();
-        return chartPanel;
-    }
-
-    /**
-     * Adds a new write mark to the series and updates the corresponding charts.
-     * This method adds the bandwidth and average data for a write mark to the respective series.
-     * If configured to show maximum and minimum values, it also updates the corresponding series.
-     * After updating the data, it refreshes the write metrics displayed on the main frame.
-     * Finally, it prints the details of the added mark to the console.
-     *
-     * @param mark The DiskMark object representing the write mark to be added.
-     */
-    public static void addWriteMark(DiskMark mark) {
-        wSeries.add(mark.getMarkNum(), mark.getBwMbSec());
-        wAvgSeries.add(mark.getMarkNum(), mark.getCumAvg());
-        if (App.showMaxMin) {
-            wMaxSeries.add(mark.getMarkNum(), mark.getCumMax());
-            wMinSeries.add(mark.getMarkNum(), mark.getCumMin());
-        }
-        Gui.mainFrame.refreshWriteMetrics();
-        System.out.println(mark.toString());
-    }
-
-    /**
-     * Adds a new read mark to the series and updates the corresponding charts.
-     * This method adds the bandwidth and average data for a read mark to the respective series.
-     * If configured to show maximum and minimum values, it also updates the corresponding series.
-     * After updating the data, it refreshes the read metrics displayed on the main frame.
-     * Finally, it prints the details of the added mark to the console.
-     *
-     * @param mark The DiskMark object representing the read mark to be added.
-     */
-    public static void addReadMark(DiskMark mark) {
-        rSeries.add(mark.getMarkNum(), mark.getBwMbSec());
-        rAvgSeries.add(mark.getMarkNum(), mark.getCumAvg());
-        if (App.showMaxMin) {
-            rMaxSeries.add(mark.getMarkNum(), mark.getCumMax());
-            rMinSeries.add(mark.getMarkNum(), mark.getCumMin());
-        }
-        Gui.mainFrame.refreshReadMetrics();
-        System.out.println(mark.toString());
-    }
-
-    /**
-     * Resets/clears the test data including the series and status bar.
-     */
-    public static void resetTestData() {
-        wSeries.clear();
-        rSeries.clear();
-        wAvgSeries.clear();
-        rAvgSeries.clear();
-        wMaxSeries.clear();
-        rMaxSeries.clear();
-        wMinSeries.clear();
-        rMinSeries.clear();
-        progressBar.setValue(0);
-        Gui.mainFrame.refreshReadMetrics();
-        Gui.mainFrame.refreshWriteMetrics();
-    }
-
-    public static void updateLegend() {
-        chart.getXYPlot().getRenderer().setSeriesVisibleInLegend(0, App.writeTest);
-        chart.getXYPlot().getRenderer().setSeriesVisibleInLegend(1, App.writeTest);
-        chart.getXYPlot().getRenderer().setSeriesVisibleInLegend(2, App.writeTest && App.showMaxMin);
-        chart.getXYPlot().getRenderer().setSeriesVisibleInLegend(3, App.writeTest && App.showMaxMin);
-
-        chart.getXYPlot().getRenderer().setSeriesVisibleInLegend(4, App.readTest);
-        chart.getXYPlot().getRenderer().setSeriesVisibleInLegend(5, App.readTest);
-        chart.getXYPlot().getRenderer().setSeriesVisibleInLegend(6, App.readTest && App.showMaxMin);
-        chart.getXYPlot().getRenderer().setSeriesVisibleInLegend(7, App.readTest && App.showMaxMin);
-    }
-}
